tests: Move subdim tests from test_yask to test_dimension
