+++ tests/test_yask.py
@@ -6,7 +6,7 @@ import pytest  # noqa
 pexpect = pytest.importorskip('yask')  # Run only if YASK is available
 
 from devito import (Eq, Grid, Operator, Constant, Function, TimeFunction,
-                    SparseTimeFunction, configuration, clear_cache, SubDimension)  # noqa
+                    SparseTimeFunction, configuration, clear_cache)  # noqa
 from devito.ir.iet import retrieve_iteration_tree  # noqa
 from devito.yask.wrappers import contexts  # noqa
 
@@ -317,145 +317,6 @@ class TestOperatorSimple(object):
         op(t=0, yu4D=u)
         assert np.all(u.data[1] == 1.)
         assert u.data[:].sum() == np.prod(grid.shape)
-
-    def test_subdims_parallel(self):
-        """
-        Tests application of an Operator consisting of a subdimension
-        defined over different sub-regions, explicitly created through the
-        use of :class:`SubDimension`s.
-        """
-        grid = Grid(shape=(20, 20))
-        x, y = grid.dimensions
-        t = grid.stepping_dim
-        thickness = 4
-
-        u = TimeFunction(name='u', save=None, grid=grid, space_order=0, time_order=1)
-
-        xi = SubDimension.middle(name='xi', parent=x,
-                                 thickness_left=thickness, thickness_right=thickness)
-
-        yi = SubDimension.middle(name='yi', parent=y,
-                                 thickness_left=thickness, thickness_right=thickness)
-
-        # a 5 point stencil that can be computed in parallel
-        centre = Eq(u[t+1, xi, yi],
-                         u[t, xi, yi] + u[t, xi-1, yi]
-                         + u[t, xi+1, yi] + u[t, xi, yi-1] + u[t, xi, yi+1])
-
-        u.data[0, 10, 10] = 1.0
-
-        op = Operator([centre])
-        op.apply(time_m=0, time_M=0)
-
-        assert np.all(u.data[1, 9:12, 10] == 1.0)
-        assert np.all(u.data[1, 10, 9:12] == 1.0)
-
-        # Other than those, it should all be 0
-        u.data[1, 9:12, 10] = 0.0
-        u.data[1, 10, 9:12] = 0.0
-        assert np.all(u.data[1, :] == 0)
-
-    def test_subdimmiddle_notparallel(self):
-        """
-        Tests application of an Operator consisting of a subdimension
-        defined over different sub-regions, explicitly created through the
-        use of :class:`SubDimension`s.
-
-        Different from case above because the interior cannot be evaluated
-        in parallel.  This should prevent yaskification.
-        """
-        grid = Grid(shape=(20, 20))
-        x, y = grid.dimensions
-        t = grid.stepping_dim
-        thickness = 4
-
-        u = TimeFunction(name='u', save=None, grid=grid, space_order=0, time_order=1)
-
-        xi = SubDimension.middle(name='xi', parent=x,
-                                 thickness_left=thickness, thickness_right=thickness)
-
-        yi = SubDimension.middle(name='yi', parent=y,
-                                 thickness_left=thickness, thickness_right=thickness)
-
-        # flow dependencies in x and y which should force serial execution
-        # in reverse direction
-        centre = Eq(u[t+1, xi, yi], u[t, xi, yi] + u[t+1, xi+1, yi+1])
-        u.data[0, 10, 10] = 1.0
-
-        op = Operator([centre])
-        op.apply(time_m=0, time_M=0)
-
-        print(u.data[1, 0:11, 0:11])
-        for i in range(4, 11):
-            assert u.data[1, i, i] == 1.0
-            u.data[1, i, i] = 0.0
-
-        assert np.all(u.data[1, :] == 0)
-
-    def test_subdimleft_notparallel(self):
-        """
-        Tests application of an Operator consisting of a subdimension
-        defined over different sub-regions, explicitly created through the
-        use of :class:`SubDimension`s.
-
-        This tests that flow direction is not being automatically inferred
-        from whether the subdimension is on the left or right boundary.
-        """
-        grid = Grid(shape=(20, 20))
-        x, y = grid.dimensions
-        t = grid.stepping_dim
-        thickness = 4
-
-        u = TimeFunction(name='u', save=None, grid=grid, space_order=1, time_order=1)
-
-        xl = SubDimension.left(name='xl', parent=x, thickness=thickness)
-
-        yi = SubDimension.middle(name='yi', parent=y,
-                                 thickness_left=thickness, thickness_right=thickness)
-
-        # Flows inward (i.e. forward) rather than outward
-        eq = Eq(u[t+1, xl, yi], u[t+1, xl-1, yi] + 1)
-
-        op = Operator([eq])
-        op.apply(time_m=0, time_M=0)
-
-        print(u.data[1, :, 10])
-        assert np.all(u.data[1, 0:4, 10] == [1, 2, 3, 4])
-        assert np.all(u.data[1, 5:, 10] == 0)
-        assert np.all(u.data[1, :, 0:10] == 0)
-        assert np.all(u.data[1, :, 11:] == 0)
-
-    def test_subdimleft_parallel(self):
-        """
-        Tests application of an Operator consisting of a subdimension
-        defined over different sub-regions, explicitly created through the
-        use of :class:`SubDimension`s.
-
-        This tests that flow direction is not being automatically inferred
-        from whether the subdimension is on the left or right boundary.
-        """
-        grid = Grid(shape=(20, 20))
-        x, y = grid.dimensions
-        t = grid.stepping_dim
-        thickness = 4
-
-        u = TimeFunction(name='u', save=None, grid=grid, space_order=0, time_order=1)
-
-        xl = SubDimension.left(name='xl', parent=x, thickness=thickness)
-
-        yi = SubDimension.middle(name='yi', parent=y,
-                                 thickness_left=thickness, thickness_right=thickness)
-
-        # Can be done in parallel
-        eq = Eq(u[t+1, xl, yi], u[t, xl, yi] + 1)
-
-        op = Operator([eq])
-        op.apply(time_m=0, time_M=0)
-
-        assert np.all(u.data[1, 0:4, 0:thickness] == 0)
-        assert np.all(u.data[1, 0:4, -thickness:] == 0)
-        assert np.all(u.data[1, 0:4, thickness:-thickness] == 1)
-        assert np.all(u.data[1, 5:, :] == 0)
 
 
 class TestOperatorAcoustic(object):
